Merge scoring code that relaxes to use ERE heads and Stanford CoreNLP heads

Adds scoring using ERE head annotation, Stanford CoreNLP parses, and some traditional head finding rules.

See merge request !15

+++ kbp-events2014-scorer/src/main/java/com/bbn/bue/common/diff/FMeasureTableRenderer.java
@@ -83,7 +83,7 @@ public final class FMeasureTableRenderer {
       new Function<FMeasureCounts, Float>() {
         @Override
         public Float apply(FMeasureCounts input) {
-          return input.falsePositives() + input.falseNegatives();
+          return (float) (input.falsePositives() + input.falseNegatives());
         }
       };
